Add a shortcut to open Todoist and focus on the new task editor.
The extension first checks if chrome already has a tab open. If there
is one, it focuses on it first. The logic for opening the editor uses
setInterval for periodic checking to see if Todoist is fully loaded.
This should be inproved upon (find a proper event to listen for).